chore(Front Generator): fix native and web generators code duplication
affects: @cuba-platform/front-generator
