+++ packages/front-generator/src/generators/common/info.json
@@ -1,0 +1,4 @@
+{
+  "clientBaseTech": "React",
+  "bower": false
+}
