+++ packages/front-generator/src/generators/react-typescript/app/index.ts
@@ -1,12 +1,10 @@
 import {ProjectInfo} from "../../../common/model/cuba-model";
-import {BaseGenerator} from "../../../common/base-generator";
-import {appGenerationOptions, CommonGenerationOptions, OptionsConfig} from "../../../common/cli-options";
+import {appGenerationOptions, CommonGenerationOptions} from "../../../common/cli-options";
 import * as path from "path";
-import {ownVersion} from "../../../cli";
 import {SdkAllGenerator} from "../../sdk/sdk-generator";
 import {SUPPORTED_CLIENT_LOCALES} from '../common/i18n';
-import {StudioTemplateProperty} from "../../../common/studio/studio-model";
-import {AppAnswers, appGeneratorParams} from "./params";
+import {appGeneratorParams} from "./params";
+import {BaseTsAppGenerator, createModel} from "../../common/BaseTsAppGenerator";
 
 export interface AppTemplateModel {
   title: string;
@@ -17,44 +15,21 @@ export interface AppTemplateModel {
   restClientSecret: string;
 }
 
-class ReactTSAppGenerator extends BaseGenerator<AppAnswers, AppTemplateModel, CommonGenerationOptions> {
+class ReactTSAppGenerator extends BaseTsAppGenerator<CommonGenerationOptions> {
 
-  conflicter!: { force: boolean }; // missing in typings
-  modelPath?: string;
-
-  constructor(args: string | string[], commonOptions: CommonGenerationOptions) {
-    super(args, commonOptions);
+  constructor(args: string | string[], options: CommonGenerationOptions) {
+    super(args, options);
     this.sourceRoot(path.join(__dirname, 'template'));
-    this.modelPath = this.options.model;
   }
 
   async prompting() {
     await this._obtainModelAndAnswers();
   }
 
-  protected async _obtainAnswers() {
-    if (!this.options.answers) {
-      this.log(`Please provide rest client id and secret.\
-        \nThis properties required for proper connection to rest service https://doc.cuba-platform.com/restapi-7.1/#rest_api_v2_ex_get_token.\
-        \nThis options (rest.client.id, rest.client.secret) could be found in web-app.properties of CUBA web module.\
-        \nIf no such properties exist in file (Studio version < 14), just skip questions and leave default answers (client/secret).`);
-    }
-    await super._obtainAnswers();
-  }
-
   // noinspection JSUnusedGlobalSymbols - yeoman runs all methods from class
   writing() {
     this.log(`Generating to ${this.destinationPath()}`);
-
-    if (!this.answers) {
-      throw new Error('Answers are not provided');
-    }
-    if (!this.cubaProjectModel) {
-      throw new Error('Model is not provided');
-    }
-
-    const {restClientId, restClientSecret} = this.answers;
-    this.model = createModel(this.cubaProjectModel.project, restClientId, restClientSecret);
+    this.model = createModel(this.cubaProjectModel, this.answers);
 
     let clientLocales: string[];
     const modelHasLocalesInfo = (this.model.project.locales != null);
@@ -117,30 +92,8 @@ class ReactTSAppGenerator extends BaseGenerator<AppAnswers, AppTemplateModel, Co
     this.log(`CUBA React client has been successfully generated into ${this.destinationRoot()}`);
   }
 
-  _getAvailableOptions(): OptionsConfig {
-    return appGenerationOptions;
-  }
-
-  _getParams(): StudioTemplateProperty[] {
-    return appGeneratorParams;
-  }
 }
 
-// todo move to AppBase generator
-export function createModel(project: ProjectInfo, restClientId: string, restClientSecret: string): AppTemplateModel {
-  return {
-    ownVersion,
-    title: project.name,
-    project,
-    basePath: project.modulePrefix + '-front',
-    restClientId: restClientId ? restClientId : 'client',
-    restClientSecret: restClientSecret ? normalizeSecret(restClientSecret) : 'secret'
-  };
-}
-
-export function normalizeSecret(restClientSecret: string): string {
-  return restClientSecret.replace(/{.*}/, '');
-}
 
 export const generator = ReactTSAppGenerator;
 export const options = appGenerationOptions;
