+++ packages/front-generator/src/test/init.test.ts
@@ -5,7 +5,7 @@ import * as fs from "fs";
 describe('init', function () {
   it('should collect clients', function () {
     const clients = collectClients('index.ts').sort((a, b) => {return a.name.localeCompare(b.name)});
-    expect(clients.length).to.eq(5);
+    expect(clients.length).to.eq(6);
 
     const genCountOfClient: Map<string, number> = new Map();
     genCountOfClient.set('polymer2', 9);
